+++ shared/server_utils/custom_config.js
@@ -17,6 +17,23 @@ function parseUrl(url) {
 	};
 }
 
+// create a modified copy of a url (both parsed and not parsed) from a parsedUrl
+// and updated host property
+function assembleInternalUrl(originalParsedUrl, internalHost=null) {
+	const parsedUrl = JSON.parse(JSON.stringify(originalParsedUrl));
+	if (internalHost) parsedUrl.host = internalHost;
+	const internalPortSufx =
+		(parsedUrl.secure && parsedUrl.port !== 443) || (!parsedUrl.secure && parsedUrl.port !== 80)
+			? (internalPortSufx = `:${parsedUrl.port}`)
+			: '';
+	const scheme = parsedUrl.secure ? 'https' : 'http';
+	const url = `${scheme}://${parsedUrl.host}${internalPortSufx}`;
+	return {
+		parsedUrl,
+		url
+	}
+}
+
 // Read the structured config to determine which broadcast engine we'll use, then
 // set the data needed for it.
 function selectBroadcastEngine(cfg) {
@@ -224,6 +241,15 @@ module.exports = function customConfigFunc(nativeCfg) {
 	}
 
 	// mongo
+	if (!nativeCfg.storage.mongo) nativeCfg.storage.mongo = {};
+	if (process.env.CSSVC_CFG_URL && process.env.CSSVC_CFG_URL !== nativeCfg.storage.mongo.url) {
+		console.error(
+			// TODO: this should be surfaced to the user, and/or the config
+			// should be automatically corrected.
+			`WARNING: CSSVC_CFG_URL(${process.env.CSSVC_CFG_URL} !== cfg (${nativeCfg.storage.mongo.url}))`
+		);
+		nativeCfg.storage.mongo.url = process.env.CSSVC_CFG_URL;
+	}
 	Object.assign(Cfg.storage.mongo, {
 		database: MongoUrlParser(nativeCfg.storage.mongo.url).database,
 		hintsRequired: true,
@@ -342,6 +368,12 @@ module.exports = function customConfigFunc(nativeCfg) {
 
 	// api
 	Cfg.apiServer.publicApiUrlParsed = parseUrl(Cfg.apiServer.publicApiUrl);
+	const internalApiUrl = assembleInternalUrl(
+		Cfg.apiServer.publicApiUrlParsed,
+		Cfg.apiServer.internalHost
+	);
+	Cfg.apiServer.internalApiUrlParsed = internalApiUrl.parsedUrl;
+	Cfg.apiServer.internalApiUrl = internalApiUrl.url;
 	Cfg.apiServer.assetEnvironment = process.env.CS_API_ASSET_ENV;
 	if (!Cfg.apiServer.authOrigin) {
 		Cfg.apiServer.authOrigin = `${Cfg.apiServer.publicApiUrl}/no-auth`;