+++ onprem_admin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [0.0.0] - 0000-0-0
+
+### Change
+
+- Schema modifications for internal connectivity
+	apiServer
+	- altBroadcasterHost, securePort
+	+ internalHost, internalPort
+	broadcaster
+	+ internalHost, internalPort
+
 ## [11.0.2] - 2021-5-18
 
 ### Changed
@@ -35,7 +46,6 @@
 ### Changed
 
 - Release bump to support latest API version
-
 ## [8.2.10] - 2021-4-6
 
 ### Changed
@@ -53,7 +63,6 @@
 ### Changed
 
 - Release bump to support latest API version
-
 ## [8.2.7] - 2021-3-17
 
 ### Changed